HotFix: 16-mixed

+++ main.py
@@ -165,7 +165,7 @@ def main(args):
     trainer = pl.Trainer(
         logger=my_loggers,
         accelerator="cpu" if hparams.gpus == 0 else "gpu",
-        precision=16 if hparams.gpus != 0 else 32,  # CPU에서는 32-bit precision
+        precision='16-mixed' if hparams.gpus != 0 else 32,  # CPU에서는 32-bit precision
         devices=None if hparams.gpus == 0 else hparams.gpus,
         callbacks=checkpoint_callback,  # 콜백 리스트로 묶는 것이 좋음
         max_epochs=hparams.epochs,
